+++ R/exdqlmPlot.R
@@ -2,7 +2,6 @@
 #'
 #' The function plots the MAP estimates and 95% credible intervals (CrIs) of the dynamic quantile of an exDQLM.
 #'
-#' @inheritParams exdqlmISVB
 #' @param m1 An object of class "\code{exdqlmMCMC}" or "\code{exdqlmISVB}".
 #' @param add If `TRUE`, the dynamic quantile will be added to existing plot.
 #' @param col Color of dynamic quantile to be plotted. Default is `purple`.
@@ -22,20 +21,17 @@
 #' model = polytrendMod(1,quantile(y,0.85),10)
 #' M0 = exdqlmISVB(y,p0=0.85,model,df=c(0.98),dim.df = c(1),
 #'                    gam.init=-3.5,sig.init=15)
-#' exdqlmPlot(y,M0,col="blue")
+#' exdqlmPlot(M0,col="blue")
 #' }
 #'
-exdqlmPlot <- function(y,m1,add=FALSE,col="purple",cr.percent=0.95){
+exdqlmPlot <- function(m1,add=FALSE,col="purple",cr.percent=0.95){
 
   # check inputs
-  check_ts(y)
   if(!is.exdqlmMCMC(m1) && !is.exdqlmISVB(m1)){
     stop("m1 must be an output from 'exdqlmISVB()' or 'exdqlmMCMC()'")
-    }
+  }
+  y = m1$y
   TT = length(y)
-  if(dim(m1$samp.theta)[2] != TT){
-    stop("length of dynamic quantile does not match data")
-    }
   p = dim(m1$samp.theta)[1]
   n.samp = dim(m1$samp.theta)[3]
   if(cr.percent<=0 | cr.percent>=1){